Board.java
Sets up the actual board for the dots game. Contains implementation of MouseListener and holds the walls and boxes (both additional classes) needed for the game.  Contains methods for the actual playing of the game, i.e. gathers (x,y) coordinates for a mouse click and implements the changing of a wall's color to that of human or computer, while also setting the region selected to be off limits after a click. Sends this information back to the dots class to allow updating of scores and turns.