Box.java
Sets up implementation of boxes, which are contained within the board class and contain four walls each, (N,S,E,W). Has methods to check which region of the box was clicked in order to decide which wall to give to the player, along with methods to set the owner of each wall/box.